Revert "Dev/First demo"
