+++ src/App.js
@@ -1,0 +1,24 @@
+import logo from './logo.svg';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Proximamente en su tienda más cercana xF 🥳
+        </p>
+        <a
+          className="App-link"
+          href="#"
+          rel="noopener noreferrer"
+        >
+          GIStura
+        </a>
+      </header>
+    </div>
+  );
+}
+
+export default App;
