--- src/App.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import { BrowserRouter, Route } from 'react-router-dom';
-import logo from './logo.svg';
-import './App.css';
-import Main from './components/MainComponent';
-
-function App() {
-  return (
-    <BrowserRouter>
-      <div>
-        <Main />
-      </div>
-    </BrowserRouter>
-  );
-}
-
-export default App;
