--- src/ccolors.css
@@ -1,3 +1,0 @@
-.orange {
-    color: orange
-}
